Make scrollable containers islands, moving the scrollable island wrapper to decide container
