+++ dotcom-rendering/src/components/ScrollableCarousel.tsx
